Assignment 1
In this assignment we have created a hill climber algorithm. A parent
vector is created with random numbers in range [0,1].
The fitness is deemed the average of the values in the parent.
With a probability of p = .05, each value of the parent vector is
replaced with another random in range [0,1], creating a new child vector
The parent and child are then compared and the vector with the highest
fitness becomes the new parent
Fig A shows the fitness values of the parent over 5000 runs
Fig B shows the fitness values of the parent over 5000 runs, run 5
times times, with each new parent vector starting from a new random
matrix
Fig C shows the individual random values of each parent plotted for
5000 runs, with the highest fitness values being the least gray