Assignment 4
In this assignment you will download, install and make some changes to
the Bullet Physics Library, the open source physics engine we will be
using for the remainder of this class. After installation and
compilation, you will run a test application that comes with Bullet
which simulates ragdolls (Fig. 1a). You will then change the radius and
length of the head capsule in the Bullet code, re-compile and re-run
the code to produce Fig. 1b. You will then turn off the physics
simulation while still creating the objects and drawing them as shwon
in Fig. 1c. You will then comment out the code that creates the
ragdolls in this example program, producing the ‘empty world’ shown in
Fig. 1d. This will provide you with a blank canvas on which you will
begin to build your robot in the next assignment.