Assignment 10.3
Added two legs, with four body parts and 4 motors/hinges, as well as
two touch sensors, to my robot to create a hexapod. Evolved the robot
for 15 or so generations to achieve high fitness moving into the
screen. Cleaned up the code significantly— my next move is to clean the
code up for my other two robots to make working on them this final week
much easier